Move data useEffects from Maker to Layout since they arent used in Maker
